Added checkpointing back, and add amp castable functions for self-attention functions

+++ onmt/modules/sinusoidal_positional_encoding.py
@@ -18,7 +18,7 @@ class SinusoidalPositionalEmbedding(nn.Module):
         :param pos_seq: sequences of RELATIVE position indices (can be negative for future)
         :param sin_first: in Attention is all you need paper, sin is first then cosin
         """
-        sinusoid_inp = torch.ger(pos_seq, self.inv_freq)
+        sinusoid_inp = torch.ger(pos_seq, self.inv_freq.type_as(pos_seq))
 
         if sin_first:
             pos_emb = torch.cat([sinusoid_inp.sin(), sinusoid_inp.cos()], dim=-1)
